% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusR.R
\name{accuracy}
\alias{accuracy}
\title{evaluate the results from the bicluster method}
\usage{
accuracy(bic.out, bic.index, bic.type)
}
\arguments{
\item{bic.out}{the biclusters obtained from the bicluster method (list).}

\item{bic.index}{the true position (row and column) of each bicluster (list).}

\item{bic.type}{the type of each bicluster (vector).}
}
\value{
a list of information about the matched true group and bicluster.
}
\description{
Function for evaluating the results from the bicluster method.
}
\examples{
\dontrun{
library(biclust)

xdim <- c(300, 600)
bicdim <- list(row=c(20,40,70), col=c(50,70,100))
bic.index <- list(row=list(1:20,21:60,61:130),
                  col=list(1:50,51:120,121:220))
type <- rep("constant",3)
M <- rep(5,3)
Me <- rep(1.5,3)
x <- bicstr("mono.rc", type, xdim, bicdim, M, Me=Me)
rownames(x) <- paste0("Cell",1:nrow(x))
colnames(x) <- paste0("Gene",1:ncol(x))

res <- biclust::biclust(x, method=BCCC())
if(res@Number==0) bic.out <- NA
if(res@Number!=0){
  bic.out <- list()
  for(jj in 1:res@Number){
    bic <- biclust::bicluster(x, res)[[jj]]
    rr <- as.numeric(na.omit(as.numeric(unlist(strsplit(rownames(bic), split="Cell")))))
    cc <- as.numeric(na.omit(as.numeric(unlist(strsplit(colnames(bic), split="Gene")))))
    bic.out[["row"]][[paste0("clus",jj)]] <- rr
    bic.out[["col"]][[paste0("clus",jj)]] <- cc
  }
 }
accuracy(bic.out, bic.index, type)
}
}
