% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiclusR.R
\name{bictype}
\alias{bictype}
\title{Generating a single bicluster}
\usage{
bictype(type, a, b, M, Mc = NA, Me = NA)
}
\arguments{
\item{type}{a character string indicating the type of bicluster and must be one of "constant", "add", "multi" and "linear" (character).}

\item{a}{the number of rows in a bicluster (integer).}

\item{b}{the number of columns in a bicluster (integer).}

\item{M}{the amplified multiple of value (numeric).}

\item{Mc}{the amplified multiple of difference for add model (numeric).}

\item{Me}{the amplified multiple of error (numeric).}
}
\value{
a bicluster matrix with dimension (a,b).
}
\description{
Function for generating a single bicluster.
}
\details{
This function can generate four type bicluster. If \code{type} is "\code{constant}", it will output a bicluster with same value. If \code{type} is "\code{add}", it will output a bicluster with a constant difference between each sample. If \code{type} is "\code{multi}", it will output a bicluster with a constant multiple between each sample. If \code{type} is "\code{linear}", it will output a bicluster with a linear relationship between samples.
}
\examples{
\dontrun{
bictype(type="constant", a = 20, b = 50, M = 5, Me = NA)
}
}
