% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CrwM}
\alias{CrwM}
\title{A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp.}
\usage{
CrwM(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of the increment}

\item{x0}{the initial value}

\item{N}{sample size}
}
\value{
a random sample of size \code{N} for the standard Laplace distribution
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp.
}
\examples{
\dontrun{
CrwM(sigma= 2, x0 = 25, N = 2000)
}
}
