\name{StatComp20004-package}
\alias{StatComp20004-package}
\alias{StatComp20004}
\docType{package}
\title{
  Functions for generating biclustering
}
\description{
  R fucntions about generating biclustering sample of different types under different structures.
}
\details{
  This package contains one function for generating a single type biclustering and one function generating biclustering sample under different structures.
}
\author{
Fengrong Liu.

Maintainer: Fengrong Liu <Liufr98@163.com>
}
\keyword{ package }

